Removed more starter code and updated filename of context
