+++ client/src/App.tsx
@@ -4,7 +4,7 @@ import { BrowserRouter, Route, Redirect, Switch } from 'react-router-dom';
 import Login from './pages/Login/Login';
 import Signup from './pages/SignUp/SignUp';
 import Dashboard from './pages/Dashboard/Dashboard';
-import { AuthProvider } from './context/AuthContext';
+import { AuthProvider } from './context/useAuthContext';
 
 import ProfileSettings from './pages/Onboarding/ProfileSettings/ProfileSettings';
 import Confirm from './pages/Onboarding/Confirm/Confirm';
