--- client/src/mocks/mockUseAuthProvider.tsx
@@ -1,19 +1,0 @@
-import { FunctionComponent } from 'react';
-import { AuthContext } from '../context/useAuthContext';
-import { mockLoggedInUser } from './mockUser';
-
-const MockUseAuthProvider: FunctionComponent = ({ children }) => {
-  return (
-    <AuthContext.Provider
-      value={{
-        loggedInUser: mockLoggedInUser,
-        updateLoginContext: jest.fn(),
-        logout: jest.fn(),
-      }}
-    >
-      {children}
-    </AuthContext.Provider>
-  );
-};
-
-export default MockUseAuthProvider;
